CLI.Parsing.Internal

Description

Internal module for CLI parsing.

Documentation

data Acc

Monoid accumulator for CLI

Constructors

Acc

Fields

accLegend :: Maybe FilePath

Path to the legend file.

accTimeout :: Maybe (RNonNegative Int)

Maximum time to run commands

accCommands :: [Text]

List of commands to run.

mapStrToEnv :: [Text] -> Maybe (RNonNegative Int) -> String -> Either ParseErr Env

Transforms a list of commands and String into Right Env. Each non-empty, non-comment (comments start with #) line in the map string are expected to have the form key=val. If any parse errors are encountered then Left ParseErr is returned.

pureParseArgs :: [String] -> ParseAnd Acc

Parses arguments into ParseAnd acc.