Git.FastForward.Types.LocalBranches

Description

Provides LocalBranches type.

Documentation

type CurrentBranch = Name

Alias for Name.

data LocalBranches

Describes the local branches in a git directory.

Constructors

LocalBranches

Fields

current :: CurrentBranch

The current branch, saved so that we can check it out after we are done updating all branches.

branches :: [Name]

List of all branches.