Git.Stale.Core.Internal

Description

Exports utility functions mainly for parsing/transforming between types/errors.

Documentation

badBranch :: Text -> Bool

Tests for bad branches based on presence of * and >.

exceptToErr :: Show a => Either a (Either Err b) -> Either Err b

Joins nested Eithers.

parseLog :: NameLog -> ErrOr NameAuthDay

Parses NameLog into NameAuthDay, recording errors as ErrOr.

parseAuthDateStr :: NameLog -> ErrOr NameAuthDateStr

Intermediate parsing of NameLog into NameAuthDateStr, recording errors as ErrOr.

parseDay :: NameAuthDateStr -> ErrOr NameAuthDay

Intermediate parsing of NameAuthDateStr into NameAuthDay, recording errors as ErrOr.

safeRead :: Text -> ErrOr Int

Safely reads Text into ErrOr Int.

stale :: RNonNegative Int -> Day -> NameAuthDay -> Bool

Determines if NameAuthDay is stale given by

stale lim day (_, _, d) <=> day - d >= lim

staleNonErr :: RNonNegative Int -> Day -> ErrOr NameAuthDay -> Bool

For Right NameAuthDay, behaves the same as stale. But for Left Err it is always true, since we do not want to filter out errors.

textToName :: Text -> ErrOr Name

Strips whitespace and potential git prefix (i.e. remotes/). Works for strings with no prefix and those that have the prefix exactly once, e.g.

textToName "branch" -> branch
textToName "remotes/origin/branch" -> origin/branch
textToName "blah/remotes/origin/branch" -> origin/branch

Returns Left ParseName if the prefix occurs more than once as we have definitely entered undefined territory.

unsafeToInt :: Text -> Int

Unsafely reads Text to Int.