Git.Stale.Parsing

Description

Handles parsing of String args into Env.

Documentation

parseArgs :: Day -> [String] -> Either ParseErr Env

Maps Day and parsed [String] args into Right Env, returning any errors as Left ParseErr. All arguments are optional (i.e. an empty list is valid), but if any are provided then they must be valid or an error will be returned. Valid arguments are:

  --grep=<string>
      Used for filtering on branch names. Any String is fine,
      defaults to the empty string.

  --path=<string>
      Path to the git directory. Any String is fine, defaults
      to the empty string (current directory).

  --limit=<days>
      Determines if a branch should be considered stale. Must be a
      non-negative integer. Defaults to 30.

  -a, --branch-type=all
      Searches local and remote branches.

  -r, --branch-type=remote
      Searches remote branches only. This is the default.

  -l, --branch-type=local
      Searches local branches only.

  --remote=<string>
      Name of the remote, used for stripping out the the remote name for
      display purposes. Any String is fine, including the empty string.
      Defaults to origin/.

  --master=<string>
      Name of the branch to consider merges against. Any String is fine,
      including the empty string. Defaults to origin/master.

 -h, --help
      Returns instructions as Left String.