Git.Stale.Types.Env

Description

Provides Env type.

Documentation

data BranchType

Describes the branch type.

Constructors

All

Remote

Local

data Env

The Env type to be used with Reader.

Constructors

Env

Fields

grepStr :: Maybe Text

A String to filter branch names on. Just s if non-empty, Nothing otherwise.

path :: Maybe FilePath

The path of the git directory. Just s if non-empty, Nothing otherwise.

limit :: NonNegative Int

A non-negative integer descrbing stale threshold in days.

branchType :: BranchType

The type of branches to search.

remoteName :: Text

The name of the remote.

master :: Text

The name of the branch to consider merges against.

today :: Day

Today's date.

branchTypeToArg :: BranchType -> String

Maps a BranchType to a String flag to be used in Core.MonadFindBranches' branchNamesByGrep command.