Git.Stale.Types.ResultsWithErrs

Description

Provides a ResultsWithErrs type that wraps Results along with a list of errors.

Documentation

newtype ErrDisp

Newtype wrapper for errors over (display string, num branches).

Constructors

ErrDisp (Text, Int)

data ResultsWithErrs

Wraps Results and includes [Err].

Constructors

ResultsWithErrs

Fields

errList :: [Err]

All errors encountered.

results :: Results

The Results.

newtype ResultsWithErrsDisp

Newtype wrapper over (ErrDisp, ResultsDisp).

Constructors

ResultsWithErrsDisp (ErrDisp, ResultsDisp)

toResultsErrDisp :: Text -> ResultsWithErrs -> ResultsWithErrsDisp

Transforms the ResultsWithErrs into a ResultsWithErrsDisp for display purposes. Strips out the prefix from the branches, if it exists.

toResultsWithErrs :: [ErrOr AnyBranch] -> ResultsWithErrs

Maps [ErrOr AnyBranch] to ResultsWithErrs.